refactor: split more parts from checker
