+++ teal/checker/file_checker.tl
@@ -1,0 +1,36 @@
+local string_checker = require("teal.checker.string_checker")
+
+local environment = require("teal.environment")
+local type Env = environment.Env
+local type Result = environment.Result
+
+local util = require("teal.util")
+local read_file_skipping_bom = util.read_file_skipping_bom
+
+local record file_checker
+end
+
+function file_checker.check(filename: string, env?: Env, fd?: FILE): Result, string
+   if env and env.loaded and env.loaded[filename] then
+      return env.loaded[filename]
+   end
+
+   local input, err: string, string
+
+   if not fd then
+      fd, err = io.open(filename, "rb")
+      if not fd then
+         return nil, "could not open " .. filename .. ": " .. err
+      end
+   end
+
+   input, err = read_file_skipping_bom(fd)
+   fd:close()
+   if not input then
+      return nil, "could not read " .. filename .. ": " .. err
+   end
+
+   return string_checker.check(input, env, filename)
+end
+
+return file_checker
